wip: separate nav to reduce async dep
