+++ src/routes/app/[conferenceId]/[committeeId]/(chairs)/+layout.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
 	import { type Snippet } from 'svelte';
+	import { goto } from '$app/navigation';
+	import { resolve } from '$app/paths';
 	import { page } from '$app/state';
 	import { client } from '$lib/api/rumbleClient/client';
 	import ChairNavbar from './ChairNavbar.svelte';
@@ -22,6 +24,7 @@
 	let { children }: Props = $props();
 
 	const committeeId = page.params.committeeId!;
+	const conferenceId = page.params.conferenceId!;
 
 	const committee = await client.liveQuery.committee({
 		__args: { id: committeeId },
@@ -126,6 +129,137 @@
 		}
 	});
 
+	const activeDraftResolutionId = $derived(committee?.activeDraftResolutionId ?? null);
+
+	const dockItems = $derived([
+		{
+			icon: 'fa-gears',
+			label: () => m.setup(),
+			href: resolve('/app/[conferenceId]/[committeeId]/(chairs)/setup', {
+				conferenceId,
+				committeeId
+			}),
+			key: 'setup'
+		},
+		{
+			icon: 'fa-users',
+			label: () => m.presence(),
+			href: resolve('/app/[conferenceId]/[committeeId]/(chairs)/presence', {
+				conferenceId,
+				committeeId
+			}),
+			key: 'presence'
+		},
+		{
+			icon: 'fa-podium',
+			label: () => m.speakersList(),
+			href: resolve('/app/[conferenceId]/[committeeId]/(chairs)/speakers-list', {
+				conferenceId,
+				committeeId
+			}),
+			key: 'speakers-list'
+		},
+		{
+			icon: 'fa-box-ballot',
+			label: () => m.voting(),
+			href: resolve('/app/[conferenceId]/[committeeId]/(chairs)/voting', {
+				conferenceId,
+				committeeId
+			}),
+			key: 'voting'
+		},
+		{
+			icon: 'fa-scroll',
+			label: () => m.resolutions(),
+			href: resolve('/app/[conferenceId]/[committeeId]/(chairs)/resolutions', {
+				conferenceId,
+				committeeId
+			}),
+			key: 'resolutions'
+		},
+		...(activeDraftResolutionId
+			? [
+					{
+						icon: 'fa-file-lines',
+						label: () => m.activeDraftResolution(),
+						href: resolve('/app/[conferenceId]/[committeeId]/(chairs)/resolutions/[paperId]', {
+							conferenceId,
+							committeeId,
+							paperId: activeDraftResolutionId
+						}),
+						key: activeDraftResolutionId
+					}
+				]
+			: [])
+	]);
+
+	function isActive(key: string) {
+		if (activeDraftResolutionId && page.params.paperId === activeDraftResolutionId) {
+			return key === activeDraftResolutionId;
+		}
+		return page.route.id?.includes(key) ?? false;
+	}
+
+	$effect(() => {
+		hotkeys('alt+1, alt+2, alt+3, alt+4, alt+5, alt+6', (event, handler) => {
+			event.preventDefault();
+			switch (handler.key) {
+				case 'alt+1':
+					goto(
+						resolve('/app/[conferenceId]/[committeeId]/(chairs)/setup', {
+							conferenceId,
+							committeeId
+						})
+					);
+					break;
+				case 'alt+2':
+					goto(
+						resolve('/app/[conferenceId]/[committeeId]/(chairs)/presence', {
+							conferenceId,
+							committeeId
+						})
+					);
+					break;
+				case 'alt+3':
+					goto(
+						resolve('/app/[conferenceId]/[committeeId]/(chairs)/speakers-list', {
+							conferenceId,
+							committeeId
+						})
+					);
+					break;
+				case 'alt+4':
+					goto(
+						resolve('/app/[conferenceId]/[committeeId]/(chairs)/voting', {
+							conferenceId,
+							committeeId
+						})
+					);
+					break;
+				case 'alt+5':
+					goto(
+						resolve('/app/[conferenceId]/[committeeId]/(chairs)/resolutions', {
+							conferenceId,
+							committeeId
+						})
+					);
+					break;
+				case 'alt+6':
+					if (activeDraftResolutionId) {
+						goto(
+							resolve('/app/[conferenceId]/[committeeId]/(chairs)/resolutions/[paperId]', {
+								conferenceId,
+								committeeId,
+								paperId: activeDraftResolutionId
+							})
+						);
+					}
+					break;
+			}
+		});
+		return () => hotkeys.unbind('alt+1, alt+2, alt+3, alt+4, alt+5, alt+6');
+	});
+
 	let committeeStatusExpiredAlerted = $state(false);
 	let speakersListOvertimeAlerted = $state(false);
 	let commentListOvertimeAlerted = $state(false);
@@ -193,11 +327,7 @@
 	<title>{committee?.abbreviation ?? 'N/A'} {m.chairControls()} - MUNify CHASE</title>
 </svelte:head>
 
-<ChairNavbar
-	title={committee?.abbreviation}
-	conferenceTitle={committee?.conference?.title}
-	activeDraftResolutionId={committee?.activeDraftResolutionId}
-/>
+<ChairNavbar title={committee?.abbreviation} conferenceTitle={committee?.conference?.title} />
 
 <div class="pb-16">
 	{@render children()}
@@ -222,3 +352,17 @@
 	committeeName={committee?.name ?? m.unknown()}
 	confettiDurationSec={20}
 />
+
+<!-- Bottom dock -->
+<div class="dock dock-md lg:dock-lg md:justify-center md:gap-4">
+	{#each dockItems as item, i (item.key)}
+		<a href={item.href} class="group relative {isActive(item.key) ? 'dock-active' : ''}">
+			<i class="fa-duotone {item.icon} size-[1.2em]"></i>
+			<span class="dock-label">{item.label()}</span>
+			<kbd
+				class="kbd kbd-sm absolute top-1/3 left-1/2 -translate-x-1/2 -translate-y-1/2 opacity-0 group-hover:opacity-100 transition-opacity backdrop-blur-sm bg-base-100/80 px-2 py-1 z-10"
+				>⌥{i + 1}</kbd
+			>
+		</a>
+	{/each}
+</div>
